fix: don't repeat menus and routes

+++ src/Generators/Scaffold/MenuGenerator.php
@@ -43,6 +43,12 @@ class MenuGenerator extends BaseGenerator
     public function generate()
     {
         $this->menuContents .= $this->menuTemplate.infy_nl();
+        $existingMenuContents = file_get_contents($this->path);
+        if (Str::contains($existingMenuContents, "<span>".$this->commandData->config->mPlural."</span>")) {
+            $this->commandData->commandObj->info('Menu '.$this->commandData->config->mPlural.' is already exists, Skipping Adjustment.');
+
+            return;
+        }
 
         file_put_contents($this->path, $this->menuContents);
         $this->commandData->commandComment("\n".$this->commandData->config->mCamelPlural.' menu added.');
